1차 장바구니,주문하기 미완성

+++ Shop_Test_KDJ/src/main/webapp/user/logout.jsp
@@ -14,23 +14,21 @@
 	cookieRememberMe.setMaxAge(0);
 	cookieToken.setMaxAge(0);
 	
-	// 세션 무효화
-	session.invalidate();
-	
 	// 쿠키 전달
 	response.addCookie(cookieRememberMe);
 	response.addCookie(cookieToken);
 	
-	
-	
 	// 아이디 가져오기
 	String loginId = (String) session.getAttribute("loginId");
-	if( loginId != null ) {
+	if( loginId != null && !loginId.equals("")) {
 		// 토큰 삭제
 		int deleted = userDAO.deleteToken(loginId);
 		if( deleted > 0 ) 	out.println("인증 토큰 데이터 삭제 성공");
 		else				out.println("인증 토큰 데이터 삭제 실패");
 	}
+	
+	// 세션 무효화
+	session.invalidate();
 	
 	// 메인 페이지로 이동
 	response.sendRedirect(request.getContextPath() + "/");
